# server/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "区块链房地产代币化系统 API",
    "version": "1.0.0",
    "description": "测试配置"
  },
  "paths": {
...
    "/api/v1/core/trading-manager/create-order": {
      "post": {
        "summary": "创建交易订单",
        "description": "创建代币交易挂单",
        "tags": [
          "TradingManager"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "token",
                  "amount",
                  "price",
                  "isSellOrder"
                ],
                "properties": {
                  "token": {
                    "type": "string",
                    "description": "代币合约地址"
                  },
                  "amount": {
                    "type": "string",
                    "description": "代币数量"
                  },
                  "price": {
                    "type": "string",
                    "description": "单价"
                  },
                  "isSellOrder": {
                    "type": "boolean",
                    "description": "是否为卖单(true)或买单(false)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "订单创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "400": {
            "description": "参数错误或合约错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/core/trading-manager/cancel-order/{orderId}": {
      "delete": {
        "summary": "取消订单",
        "description": "取消指定ID的交易订单",
        "tags": [
          "TradingManager"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "orderId",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "订单ID"
          }
        ],
        "responses": {
          "200": {
            "description": "订单取消成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "400": {
            "description": "参数错误或合约错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/core/trading-manager/execute-order/{orderId}": {
      "post": {
        "summary": "执行订单",
        "description": "执行指定ID的交易订单",
        "tags": [
          "TradingManager"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "orderId",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "订单ID"
          }
        ],
        "responses": {
          "200": {
            "description": "订单执行成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "400": {
            "description": "参数错误或合约错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/core/trading-manager/order/{orderId}": {
      "get": {
        "summary": "获取订单详情",
        "description": "获取指定ID订单的详细信息",
        "tags": [
          "TradingManager"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "orderId",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "订单ID"
          }
        ],
        "responses": {
          "200": {
            "description": "订单详情获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "404": {
            "description": "订单不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/core/trading-manager/user-orders/{address}": {
      "get": {
        "summary": "获取用户订单",
        "description": "获取指定用户的所有订单ID",
        "tags": [
          "TradingManager"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "address",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "用户地址"
          }
        ],
        "responses": {
          "200": {
            "description": "用户订单获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/core/trading-manager/trade/{tradeId}": {
      "get": {
        "summary": "获取交易详情",
        "description": "获取指定ID交易的详细信息",
        "tags": [
          "TradingManager"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "tradeId",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "交易ID"
          }
        ],
        "responses": {
          "200": {
            "description": "交易详情获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "404": {
            "description": "交易不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/core/trading-manager/user-trades/{address}": {
      "get": {
        "summary": "获取用户交易",
        "description": "获取指定用户的所有交易ID",
        "tags": [
          "TradingManager"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "address",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "用户地址"
          }
        ],
        "responses": {
          "200": {
            "description": "用户交易获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/core/trading-manager/token-trades/{tokenAddress}": {
      "get": {
        "summary": "获取代币交易",
        "description": "获取指定代币的所有交易ID",
        "tags": [
          "TradingManager"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "tokenAddress",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "代币合约地址"
          }
        ],
        "responses": {
          "200": {
            "description": "代币交易获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          }
        }
      }
    }
  },
...
}